scripts/requirements.txt
gtfs-kit
supabase
python-dotenv
